fix: address CodeRabbit review findings

+++ crates/koji-core/src/proxy/server/mod.rs
@@ -79,7 +79,7 @@ impl ProxyServer {
                         vram_total_mib: sample.vram.as_ref().map(|v| v.total_mib as i64),
                         models_loaded: sample.models_loaded as i64,
                     };
-                    let cutoff_ms = sample.ts_unix_ms - (retention_secs as i64 * 1000);
+                    let cutoff_ms = sample.ts_unix_ms - (retention_secs as i128 * 1000) as i64;
                     // Run the blocking SQLite call off the runtime.
                     let _ = tokio::task::spawn_blocking(move || {
                         if let Err(e) =
